+++ packages/orq-rc/src/orq_ai_sdk/memorystores.py
@@ -4,7 +4,6 @@ from .basesdk import BaseSDK
 from datetime import datetime
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import creatememorystoreop as models_creatememorystoreop
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -215,8 +214,8 @@ class MemoryStores(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_creatememorystoreop.CreateMemoryStoreRequestBody,
-                models_creatememorystoreop.CreateMemoryStoreRequestBodyTypedDict,
+                models.CreateMemoryStoreRequestBody,
+                models.CreateMemoryStoreRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -316,8 +315,8 @@ class MemoryStores(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_creatememorystoreop.CreateMemoryStoreRequestBody,
-                models_creatememorystoreop.CreateMemoryStoreRequestBodyTypedDict,
+                models.CreateMemoryStoreRequestBody,
+                models.CreateMemoryStoreRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,