+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -68,7 +68,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc8y89hp31vvv97sg7v848jc"
+        "conv_01kc9bp7ezd2zj41nrabed8vvk"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """