+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -254,5 +254,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-12T09:28:31.861Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-12T13:23:17.825Z")
     r"""The date and time the resource was last updated"""