+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -98,7 +98,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc8y89hrxwrw5h0c13vy02nr"
+        "conv_01kc9bp7f1q1dkaznymh79504e"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """