+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC8Y8A0BNH5AA82KKWVS67PZ"
+        "01KC9BP7W794VYM74NZJDZ1ZDT"
     )
     r"""The unique identifier of the data source"""