+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -150,7 +150,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc8y89htktfvg44ptdfeqtky"
+        "conv_01kc9bp7fb089w74125j6j0j25"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """