+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveconversationop.py
@@ -110,7 +110,7 @@ class RetrieveConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc8y89hws2pjs5e0zfttefck"
+        "conv_01kc9bp7fd095jx5h13h41hxfp"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """