+++ packages/orq-rc/RELEASES.md
@@ -4399,3 +4399,13 @@ Based on:
 - [python v4.1.0-rc.40] packages/orq-rc
 ### Releases
 - [PyPI v4.1.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.40 - packages/orq-rc
+
+## 2025-12-12 13:25:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.675.4 (2.779.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0-rc.41] packages/orq-rc
+### Releases
+- [PyPI v4.1.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.41 - packages/orq-rc